% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_box.R
\name{create_box}
\alias{create_box}
\title{Creates a unicode box this length matching text}
\usage{
create_box(text, print = FALSE)
}
\arguments{
\item{text}{A string}

\item{print}{Boolean, whether to print result or not}
}
\value{
vector of 3 lines (top - text - bottom)
}
\description{
Creates a unicode box this length matching text
}
\examples{
create_box('Input A')
}
