% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_diagram.R
\name{create_data_diagram}
\alias{create_data_diagram}
\title{Wrapper around create_data_diagram_horizontal and create_data_diagram_vertical}
\usage{
create_data_diagram(inputs, outputs, shape = "horizontal")
}
\arguments{
\item{inputs}{a vector of strings}

\item{outputs}{a vector of strings}

\item{shape}{}
}
\value{
A Unicode horizontal or vertical data diagram
}
\description{
Wrapper around create_data_diagram_horizontal and create_data_diagram_vertical
}
\examples{
create_data_diagram(inputs = c("Input A", "Input B"), outputs = c("Output A", "Output B"), shape = "horizontal")
}
