% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_diagram_horizontal.R
\name{create_data_diagram_horizontal}
\alias{create_data_diagram_horizontal}
\title{Creates an horizontal data diagram}
\usage{
create_data_diagram_horizontal(inputs, outputs)
}
\arguments{
\item{inputs}{a vector of strings}

\item{outputs}{a vector of strings}
}
\value{
A Unicode horizontal data diagram
}
\description{
Creates an horizontal data diagram
}
\examples{
create_data_diagram_horizontal(inputs = c("Input A", "Input B"), outputs = c("Output A", "Output B"))
}
