% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_diagram_vertical.R
\name{create_data_diagram_vertical}
\alias{create_data_diagram_vertical}
\title{Creates a vertical data diagram}
\usage{
create_data_diagram_vertical(inputs, outputs)
}
\arguments{
\item{inputs}{a vector of strings}

\item{outputs}{a vector of strings}
}
\value{
A Unicode horizontal data diagram
}
\description{
Creates a vertical data diagram
}
\examples{
create_data_diagram_vertical(inputs = c("Input A", "Input B"), outputs = c("Output A", "Output B"))
}
